services/predict-engine/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
mesa==2.1.5
numpy==1.26.0
ortools==9.8.3296
pandas==2.2.2
asyncpg==0.29.0
faker==25.0.0
python-dotenv==1.0.1
pydantic==2.7.0
bcrypt==4.1.3
